Remove all remaining inline test routes from index.js — fully extracted to routes/tests.js
